+++ src/new-components/banner/banner.module.css
@@ -12,14 +12,21 @@
 	background-color: #131316;
 }
 
-.banner_text {
+.banner_article {
 	display: flex;
-	flex-flow: column;
+	flex-flow: row;
 	align-items: center;
 	justify-content: center;
 	width: 100%;
-	height: 164px;
+	height: 104px;
+	text-align: start;
 	text-align: center;
+}
+
+.banner_article_inner {
+	display: block;
+	width: 100%;
+	margin-left: 32px;
 }
 
 .banner_profile {
@@ -53,6 +60,16 @@
 	align-items: end;
 	justify-content: end;
 	width: 100%;
+}
+
+.banner_text {
+	display: flex;
+	flex-flow: column;
+	align-items: center;
+	justify-content: center;
+	width: 100%;
+	height: 164px;
+	text-align: center;
 }
 
 .banner_shadow {
