+++ src/new-components/banner/banner.tsx
@@ -1,170 +1,44 @@
-import React, { useState } from "react";
-import { Redirect } from "react-router-dom";
+import React from "react";
 
-import Button from "../button/button";
-import { AvatarIcon } from "../../assets/icons";
+import { ArticleBanner, ProfileBanner, AppBanner } from "./components";
 
 import styles from "./banner.module.css";
 
-type TArticle = {
-	slug: string;
-	title: string;
-	description: string;
-	body: string;
-	createdAt: Date;
-	updatedAt: Date;
-	tagList: string[];
-	favorited: boolean;
-	favoritesCount: number;
-};
-
-type TProfile = {
-	username: string;
-	image: string;
-	following: boolean;
-};
-
-type TBannerInfo = {
+export type TBannerInfo = {
 	appName: string;
 };
 
-type TTextBanner = {
+export type TTextBanner = {
 	type: "text";
 	appName: string;
 	onFollow?: never;
 	onUnfollow?: never;
-	onEditArticle?: never;
-	onRemoveArticle?: never;
 	isUser?: never;
 	user?: never;
 	article?: never;
 };
 
-type TProfileBanner = {
+export type TProfileBanner = {
 	type: "profile";
 	appName?: never;
 	onFollow: () => void;
 	onUnfollow: () => void;
-	onEditArticle?: never;
-	onRemoveArticle?: never;
 	isUser: boolean;
-	user: TProfile;
+	user: TUser;
 	article?: never;
 };
 
-type TArticleBanner = {
+export type TArticleBanner = {
 	type: "article";
 	appName?: never;
 	onFollow?: never;
 	onUnfollow?: never;
-	onEditArticle: () => void;
-	onRemoveArticle: () => void;
 	isUser: boolean;
 	user?: never;
 	article: TArticle;
 };
 
 type TBanner = TTextBanner | TProfileBanner | TArticleBanner;
-
-function EditUserProfile({ isUser }: { isUser: boolean }) {
-	const [redirect, setRedirect] = useState(false);
-	if (!isUser) {
-		return null;
-	}
-
-	return (
-		<>
-			<Button
-				type="button"
-				onClick={() => {
-					setRedirect(true);
-				}}
-				iconType="edit"
-			>
-				Редактировать профиль
-			</Button>
-			{redirect && <Redirect to="/settings" push={true} />}
-		</>
-	);
-}
-
-function FollowUserButton({
-	isUser,
-	user,
-	onFollow,
-	onUnfollow,
-}: {
-	isUser: boolean;
-	user: TProfile;
-	onFollow: (username: string) => void;
-	onUnfollow: (username: string) => void;
-}) {
-	const handleClick = () => {
-		if (user?.following) {
-			onUnfollow(user.username);
-		} else {
-			onFollow(user.username);
-		}
-	};
-
-	if (isUser) {
-		return null;
-	}
-
-	return (
-		<Button
-			type="button"
-			onClick={handleClick}
-			iconType={user.following ? "unfollow" : "follow"}
-		>
-			{user.following ? "Unfollow" : "Follow"}
-		</Button>
-	);
-}
-
-function ProfileBanner({
-	isUser,
-	user,
-	onFollow,
-	onUnfollow,
-}: Omit<TProfileBanner, "type">) {
-	return (
-		<div className={styles["banner_profile"]}>
-			<div className={styles["banner_profile_inner"]}>
-				<AvatarIcon width={120} height={120} />
-				<h3
-					className={`${styles["banner_profile__username"]} text text_type_main-headline`}
-				>
-					{user.username}
-				</h3>
-
-				<div className=""></div>
-				<div className={styles["banner_profile_buttons"]}>
-					<EditUserProfile isUser={isUser} />
-					<FollowUserButton
-						isUser={isUser}
-						user={user}
-						onFollow={onFollow}
-						onUnfollow={onUnfollow}
-					/>
-				</div>
-			</div>
-		</div>
-	);
-}
-
-function TextBanner({ appName }: TBannerInfo) {
-	return (
-		<div className={styles["banner_text"]}>
-			<h1 className={`${styles["banner_shadow"]} text text_type_main-h1`}>
-				{appName}
-			</h1>
-			<h3 className={`${styles["banner_shadow"]} text text_type_main-headline`}>
-				Место, где готовится новый опыт
-			</h3>
-		</div>
-	);
-}
 
 export function Banner({
 	appName,
@@ -174,12 +48,11 @@ export function Banner({
 	article,
 	onFollow,
 	onUnfollow,
-	onEditArticle,
-	onRemoveArticle,
 }: TBanner) {
 	let content;
 	switch (type) {
 		case "article":
+			content = <ArticleBanner isUser={isUser!} article={article!} />;
 			break;
 		case "profile":
 			content = (
@@ -192,10 +65,10 @@ export function Banner({
 			);
 			break;
 		case "text":
-			content = <TextBanner appName={appName!} />;
+			content = <AppBanner appName={appName!} />;
 			break;
 		default:
-			content = <TextBanner appName={appName!} />;
+			content = <AppBanner appName={appName!} />;
 			break;
 	}
 	return (
