+++ src/new-components/banner/components/index.ts
@@ -1,0 +1,3 @@
+export { ArticleBanner } from "./article-banner";
+export { ProfileBanner } from "./profile-banner";
+export { AppBanner } from "./app-banner";
